+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/_params.py
@@ -281,11 +281,11 @@ def load_arguments(self, _):
         c.argument('method', arg_type=get_enum_type(['head', 'get', 'put', 'post', 'delete', 'options', 'patch'], default='get'),
                    help='HTTP request method')
         c.argument('url', help="request url")
-        c.argument('headers', nargs='+', help="Space-separated appsettings in KEY=VALUE format or JSON string. Use @{file} to load from a file")
-        c.argument('uri_parameters', nargs='+', help='Space-separated appsettings in KEY=VALUE format or JSON string. Use @{file} to load from a file')
+        c.argument('headers', nargs='+', help="Space-separated headers in KEY=VALUE format or JSON string. Use @{file} to load from a file")
+        c.argument('uri_parameters', nargs='+', help='Space-separated queries in KEY=VALUE format or JSON string. Use @{file} to load from a file')
         c.argument('skip_authorization_header', action='store_true', help='do not auto append "Authorization" header')
         c.argument('body', help='request body')
         c.argument('output_file', help='save response payload to a file')
         c.argument('resource', help='Resource which CLI would acquire token so to put on the "Authorization" header. '
                    'By default, CLI can figure this out based on --url, unless you use ones not in the list '
-                   'of "az cloud show"')
+                   'of "az cloud show". E.g. --resource https://graph.microsoft.com/')