+++ src/command_modules/azure-cli-resource/HISTORY.rst
@@ -2,8 +2,10 @@
 
 Release History
 ===============
-* 'az rest': new command for making REST calls
 
+2.1.16
+++++++
+* 'az rest': new command for making REST calls
 * `policy assignment list`: Fix error when using a resource group or subscription level `--scope`.
 
 2.1.15